@@ -14,7 +14,7 @@ description: A Tutorial App that shows how to connect network data with the BLoC
 version: 1.0.0+1
 
 environment:
-  sdk: ">=2.1.0 <3.0.0"
+  sdk: '>=2.5.2 <3.0.0'
 
 dependencies:
   flutter:
@@ -23,8 +23,9 @@ dependencies:
   # The following adds the Cupertino Icons font to your application.
   # Use with the CupertinoIcons class for iOS style icons.
   cupertino_icons: ^0.1.2
-
   http: ^0.12.0+2
+  flutter_bloc: ^4.0.0
+  equatable: ^1.0.0
 
 dev_dependencies:
   flutter_test:
