@@ -1,5 +1,6 @@
+export 'query/bloc.dart';
+export 'location/bloc.dart';
+export 'restaurant/bloc.dart';
+export 'favorite/bloc.dart';
 
-abstract class Bloc {
-  void dispose();
-}
 
