@@ -1,20 +1,42 @@
 import 'package:flutter/material.dart';
-import 'package:restaurant_finder/Bloc/bloc_provider.dart';
-import 'package:restaurant_finder/Bloc/favorite_bloc.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:restaurant_finder/Bloc/bloc.dart';
 import 'package:restaurant_finder/UI/restaurant_tile.dart';
-import '../DataLayer/restaurant.dart';
 
 class FavoriteScreen extends StatelessWidget {
 
   @override
   Widget build(BuildContext context) {
-    final bloc = BlocProvider.of<FavoriteBloc>(context);
 
     return Scaffold(
       appBar: AppBar(
         title: Text('Favorites'),
       ),
-      body: StreamBuilder<List<Restaurant>>(
+      body: BlocBuilder<FavoriteBloc, FavoriteState>(
+
+      builder: (context, state) {
+
+          if (state is FavoriteUpdated) {
+            final favorites = state.restaurants;
+
+            if (favorites == null || favorites.isEmpty) {
+              return Center(child: Text('No Favorites'));
+            }
+
+            return ListView.separated(
+              itemCount: favorites.length,
+              separatorBuilder: (context, index) => Divider(),
+              itemBuilder: (context, index) {
+                final restaurant = favorites[index];
+                return RestaurantTile(restaurant: restaurant);
+              },
+            );
+          }
+          return Center(child: Text('No Favorites'));
+        }
+      ),
+
+      /*body: StreamBuilder<List<Restaurant>>(
         stream: bloc.favoritesStream,
         // 1
         initialData: bloc.favorites,
@@ -38,7 +60,7 @@ class FavoriteScreen extends StatelessWidget {
             },
           );
         },
-      ),
+      ),*/
     );
   }
 }
