@@ -1,26 +1,18 @@
 import 'package:flutter/material.dart';
-import 'package:restaurant_finder/Bloc/bloc_provider.dart';
-import 'package:restaurant_finder/Bloc/location_bloc.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:restaurant_finder/Bloc/bloc.dart';
 import 'package:restaurant_finder/UI/restaurant_screen.dart';
-import '../DataLayer/location.dart';
 import 'location_screen.dart';
 
 class MainScreen extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return StreamBuilder<Location>(
-      stream: BlocProvider.of<LocationBloc>(context).locationStream,
-      builder: (context, snapshot) {
-        final location = snapshot.data;
-
-        if (location == null) {
-          return LocationScreen();
-        }
-
+    return BlocBuilder<LocationBloc, LocationState>(builder: (context, state) {
+      if (state is LocationPushed) {
+        final location = state.location;
         return RestaurantScreen(location: location);
-
-      },
-    );
-
+      }
+      return LocationScreen();
+    });
   }
 }
