@@ -1,18 +1,21 @@
 import 'package:flutter/material.dart';
-import 'Bloc/bloc_provider.dart';
-import 'Bloc/favorite_bloc.dart';
-import 'Bloc/location_bloc.dart';
+import 'package:flutter_bloc/flutter_bloc.dart';
+import 'package:restaurant_finder/Bloc/bloc.dart';
+import 'package:restaurant_finder/Bloc/simple_bloc_delegate.dart';
 import 'UI/main_screen.dart';
 
-void main() => runApp(RestaurantFinder());
+void main() {
+  BlocSupervisor.delegate = SimpleBlocDelegate();
+  runApp(RestaurantFinder());
+}
 
 class RestaurantFinder extends StatelessWidget {
   @override
   Widget build(BuildContext context) {
-    return BlocProvider<LocationBloc>(
-      bloc: LocationBloc(),
-      child: BlocProvider<FavoriteBloc>(
-        bloc: FavoriteBloc(),
+    return BlocProvider(
+      create: (context) => FavoriteBloc(),
+      child: BlocProvider(
+        create: (context) => LocationBloc(),
         child: MaterialApp(
           title: 'Restaurant Finder',
           theme: ThemeData(
@@ -22,6 +25,5 @@ class RestaurantFinder extends StatelessWidget {
         ),
       ),
     );
-
   }
 }
